+++ controllers/contacts/getById.js
@@ -1,0 +1,19 @@
+const contacts = require("../../models/contacts");
+const { HttpError } = require("../../helpers");
+
+const getById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contacts.getContactById(contactId);
+
+    if (!result) {
+      throw HttpError(404, "Not found");
+    }
+
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getById;
